(maint) Refactor Puppet Gatling
Prior to this commit, the artifactID was misnamed, the package name
needed to be changed, the constant class should of been class not
interface, the constants in the Constant class did not follow the java
convention for constants, and the path to the puppet.png image was wrong
due to the new package name. This commit fixes all of those things.