(feature) Finalize Puppet Gatling
This commit adds new functionality to the puppet gatling Jenkins plugin.
It redoes how it parses the gatling statistics file due to the new
changes in the gatling plugin. Adds logic to grab data from the
gatling-puppet-load-test ruby project, adds graphing and way to
calculate data based on puppet, and adds OS statistics and pie charts on
the build page.

And now begins the code review...and eventual design overhaul.