(maint) Switch FileUtils to IOUtils
Prior to this commit, puppet-gatling was using FileUtils to get the file
path from the gatling-puppet-load-test data files. This worked fine on
my local test of jenkins, however when the real plugin was added to a
legacy jenkins box, the job ran on a jenkins slave. The problem with
this is that the plugin had no idea about the jenkins slave, and the
plugin thought the files were localted on the jenkins master. This
commit fixes this problem by using IOUtils and the FilePath .read()
method to figure out where the files are located.