(maint) Fix breaking error with files not being found
Prior to this commit, the plugin would fail when you had a simulation
that used more than one version of Puppet Enterprise. This commit fixes
that by looking in the location of the workspace for the stats files
rather than within the git repository for gatling-puppet-load-test.
Another change in this project was done in gatling-puppet-load-test to
make the project work. It saved the data files outside of the git repo.