(SERVER-1501) Initial work for memory graph
This commit adds a memory graph to the main detail view for a
build.  The graph is currently populated with bunk data, and
has not yet been added to the view you see if you click the
'puppet gatling' link in the sidebar.  Needs follow-up commits
for adding it to the other view, parsing the real data from
the metrics.json, and conditioning the visibility of the
graph based on whether or not the metrics.json file is found.